Extract API client, write a simple test
The dependency-apocalypse is incoming...

We extract all-things-HTTP into an API-client, that is injected into the
TmdbService. That way we can easily mock the outcome, and just rely on
throwing exceptions to represent HTTP errors.

"supertest" seems to be the way to test express apps.

I'm using IDEA, and jest would fail to execute an X.test.ts file unless
a jest.config.js exists (generated by running
[`yarn ts-jest config:init`](https://kulshekhar.github.io/ts-jest/docs/getting-started/installation)).